fix: animation hook 분리
