+++ packages/components/src/pickers/DatePicker/hooks/useDatePicker.ts
@@ -1,11 +1,7 @@
-import { useState, useMemo, useRef } from 'react';
-import { Animated, Easing, LayoutAnimation, Platform, UIManager } from 'react-native';
+import { useState, useMemo } from 'react';
 import { useTriggerHaptic } from '../../../hooks/useTriggerHaptic';
 import { getFirstDayOfMonth, getDaysInMonth, getTodayMidnight } from '../../utils';
-
-if (Platform.OS === 'android' && UIManager.setLayoutAnimationEnabledExperimental) {
-  UIManager.setLayoutAnimationEnabledExperimental(true);
-}
+import { useCalendarAnimation } from './useCalendarAnimation';
 
 interface UseDatePickerProps {
   initialSelectedDate?: Date;
@@ -25,6 +21,15 @@ export const useDatePicker = ({
   onPress,
 }: UseDatePickerProps) => {
   const { triggerHaptic } = useTriggerHaptic();
+  const {
+    prevArrowScale,
+    nextArrowScale,
+    calendarOpacity,
+    calendarTranslateX,
+    animatePress,
+    animateMonthTransition,
+    animateDateSelect,
+  } = useCalendarAnimation();
 
   const today = useMemo(() => getTodayMidnight(), []);
 
@@ -34,67 +39,6 @@ export const useDatePicker = ({
   const [currentYear, setCurrentYear] = useState(defaultDate.getFullYear());
   const [currentMonth, setCurrentMonth] = useState(defaultDate.getMonth());
   const [selectedDate, setSelectedDate] = useState<Date>(defaultDate);
-
-  const prevArrowScale = useRef(new Animated.Value(1)).current;
-  const nextArrowScale = useRef(new Animated.Value(1)).current;
-  const calendarOpacity = useRef(new Animated.Value(1)).current;
-  const calendarTranslateX = useRef(new Animated.Value(0)).current;
-
-  const animatePress = (scale: Animated.Value, callback?: () => void) => {
-    Animated.sequence([
-      Animated.spring(scale, {
-        toValue: 0.92,
-        useNativeDriver: true,
-        speed: 50,
-        bounciness: 4,
-      }),
-      Animated.spring(scale, {
-        toValue: 1,
-        useNativeDriver: true,
-        speed: 20,
-        bounciness: 6,
-      }),
-    ]).start(callback);
-  };
-
-  const animateMonthTransition = (
-    direction: 'prev' | 'next',
-    callback: () => void
-  ) => {
-    const translateValue = direction === 'prev' ? 30 : -30;
-
-    Animated.parallel([
-      Animated.timing(calendarOpacity, {
-        toValue: 0,
-        duration: 120,
-        useNativeDriver: true,
-        easing: Easing.out(Easing.ease),
-      }),
-      Animated.timing(calendarTranslateX, {
-        toValue: translateValue,
-        duration: 120,
-        useNativeDriver: true,
-        easing: Easing.out(Easing.ease),
-      }),
-    ]).start(() => {
-      callback();
-      calendarTranslateX.setValue(-translateValue);
-      Animated.parallel([
-        Animated.timing(calendarOpacity, {
-          toValue: 1,
-          duration: 180,
-          useNativeDriver: true,
-          easing: Easing.out(Easing.ease),
-        }),
-        Animated.spring(calendarTranslateX, {
-          toValue: 0,
-          useNativeDriver: true,
-          speed: 20,
-          bounciness: 4,
-        }),
-      ]).start();
-    });
-  };
 
   const firstDay = getFirstDayOfMonth(currentYear, currentMonth);
   const daysInMonth = getDaysInMonth(currentYear, currentMonth);
@@ -149,20 +93,7 @@ export const useDatePicker = ({
     if (isDateDisabled(date)) return;
     triggerHaptic();
     
-    LayoutAnimation.configureNext({
-      duration: 250,
-      create: {
-        type: LayoutAnimation.Types.easeInEaseOut,
-        property: LayoutAnimation.Properties.opacity,
-      },
-      update: {
-        type: LayoutAnimation.Types.easeInEaseOut,
-      },
-      delete: {
-        type: LayoutAnimation.Types.easeInEaseOut,
-        property: LayoutAnimation.Properties.opacity,
-      },
-    });
+    animateDateSelect();
     
     const newDate = new Date(currentYear, currentMonth, date);
     setSelectedDate(newDate);
